+++ slurm_drmaa/job.c
@@ -432,6 +432,12 @@ static void
 slurmdrmaa_job_on_missing( fsd_job_t *self )
 {
 	job_id_spec_t job_id_spec;
+	slurmdb_job_cond_t *job_cond = NULL;
+	slurmdb_job_rec_t *job = NULL;
+	List jobs;
+	ListIterator itr = NULL;
+	char *start;
+	void *acct_db_conn = NULL;
 
 	fsd_log_enter(( "({job_id=%s})", self->job_id ));
 
@@ -447,8 +453,45 @@ slurmdrmaa_job_on_missing( fsd_job_t *self )
 		self->exit_status = 0;
 	}
 	else {
-		self->state = DRMAA_PS_FAILED; /* otherwise failed */
+		self->state = DRMAA_PS_FAILED;  /* assume failed if no job was found */
 		self->exit_status = -1;
+
+		fsd_log_info(( "Job %s has status %d, looking into accounting infos", self->job_id, self->state ));
+
+		job_cond = calloc(1, sizeof(slurmdb_job_cond_t));
+		job_cond->db_flags = SLURMDB_JOB_FLAG_NOTSET;
+		job_cond->flags |= JOBCOND_FLAG_NO_TRUNC;
+		job_cond->step_list = slurm_list_create(slurmdb_destroy_selected_step);
+
+		slurm_addto_step_list(job_cond->step_list, self->job_id);
+		job_cond->usage_end = time(NULL);
+        acct_db_conn = slurmdb_connection_get();
+		jobs = slurmdb_jobs_get(acct_db_conn, job_cond);
+		slurmdb_connection_close(&acct_db_conn);
+		slurm_list_destroy(job_cond->step_list);
+		free(job_cond);
+
+		if (jobs) {
+			itr = slurm_list_iterator_create(jobs);
+
+			job = slurm_list_next(itr);
+			if (!job) {
+				fsd_log_info(( "Job not found in accounting." ));
+			} else {
+				do {
+					fsd_log_info(( "Job found in accounting with exit code %d", job->exitcode ));
+					self->exit_status = job->exitcode;
+					if (self->exit_status == 0)
+						self->state = DRMAA_PS_DONE;
+					else
+						self->state = DRMAA_PS_FAILED;
+
+					break;
+				} while ((job = slurm_list_next(itr)));
+			}
+
+			slurm_list_iterator_destroy(itr);
+		}
 	}
 
 	fsd_log_info(("job_on_missing evaluation result: state=%d exit_status=%d", self->state, self->exit_status));