+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+HEAD (unreleased)
+-----------------
+
+### New Features and Enhancements
+
+- Retrieve job status from accounting (slurmdb) if already complete (instead of only looking at slurmd) (@holtgrewe)
+
 1.1.2 (2021-01-27)
 ------------------